pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.7",
    "langchain-openai>=1.0.3",
    "langchain-tavily>=0.2.13",
    "tavily-python>=0.7.13",
]
